isolate the compilation of reduce kernels
